Refactor Cython code, new vrt module.
